refactor/modules/moved app-routing to modules folder
