+++ src/app/modules/app.module.ts
@@ -1,7 +1,7 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 
-import { AppRoutingModule } from '../app-routing.module';
+import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from '../app.component';
 import { HeaderComponent } from '../utils/UI/header/header.component';
 import { AuthModule } from './auth/auth.module';
